# Densely pack valid block weight modes (10-20% speedup)