Since there's only a small subset of 2048 ones that are valid. For 2D LDR block sizes, this is the valid mode count -- 4x4: 145, 5x5: 255, 6x6: 370, 8x8: 555, 10x10: 687, 12x12: 775. But the code was looping over all the possible 2048 ones, doing checks "is this valid?" or sentinel values to skip over them.

Coding time of 2048x2048 image:

- 6x6 medium: Mac 4.28s -> 3.81s, PC 2.38s -> 2.22s
- 6x6 fast: Mac 0.72s -> 0.57s, PC 0.64s -> 0.57s
- 4x4 medium: Mac 8.53s -> 6.47s, PC 3.92s -> 3.28s
- 4x4 fast: Mac 1.27s -> 0.88s, PC 0.95s -> 0.76s

(Mac being 2018 MBP, Core i9 2.9GHz; PC being AMD ThreadRipper 1950X 3.4GHz; both tested with AVX2 code path)

It's entirely possible I have messed something up, e.g. missed an index update somewhere. That said, PSNR on this one image I tested on is exactly the same before & after the change.