@solidpixel 2020-10-29
Other than the name shadowing problems, which trip our compiles, the tests pass OK for me - thrown LDR, HDR, normals, and 3D at it and PSNR is identical.